requirements.txt
aws-cdk.aws-dynamodb==1.121.0
aws-cdk.aws-lambda==1.121.0
aws-cdk.aws-lambda-python==1.121.0
aws-cdk.core==1.121.0
